+++ starkinfra/staticbrcode/static_brcode.go
@@ -22,8 +22,10 @@ import (
 //
 //	Parameters (optional):
 //	- Amount [int, default 0]: Positive int that represents the amount in cents of the resulting Pix transaction. If the amount is zero, the sender can choose any amount in the moment of payment. ex: 1234 (= R$ 12.34)
-//  - ReconciliationId [string, default nil]: Id to be used for conciliation of the resulting Pix transaction. This id must have up to 25 alphanumeric digits ex: "ah27s53agj6493hjds6836v49"
-//  - Tags [slice of strings, default nil]: Slice of strings for tagging. ex: []string{"travel", "food"}
+//	- ReconciliationId [string, default nil]: Id to be used for conciliation of the resulting Pix transaction. This id must have up to 25 alphanumeric digits ex: "ah27s53agj6493hjds6836v49"
+//	- CashierBankCode [string, default nil]: Cashier's bank code. ex: "20018183".
+//	- Description [string, default nil]: Optional description to override default description to be shown in the bank statement. ex: "Payment for service #1234"
+//	- Tags [slice of strings, default nil]: Slice of strings for tagging. ex: []string{"travel", "food"}
 //
 //	Attributes (return-only):
 //	- Id [string]: Id returned on creation, this is the BR code. ex: "00020126360014br.gov.bcb.pix0114+552840092118152040000530398654040.095802BR5915Jamie Lannister6009Sao Paulo620705038566304FC6C"
@@ -38,6 +40,8 @@ type StaticBrcode struct {
 	City             string     `json:",omitempty"`
 	Amount           int        `json:",omitempty"`
 	ReconciliationId string     `json:",omitempty"`
+	CashierBankCode  string     `json:",omitempty"`
+	Description      string     `json:",omitempty"`
 	Tags             []string   `json:",omitempty"`
 	Id               string     `json:",omitempty"`
 	Uuid             string     `json:",omitempty"`